case 1516, adding observer to expire cache when content changes
git-svn-id: https://svn.browsermedia.com/bmedia/browsercms/trunk@3114 0d4fd071-7843-0410-978a-dc594407b538

+++ lib/cms/extensions/active_support/cache/file_store.rb
@@ -2,7 +2,7 @@ module ActiveSupport
   module Cache
     class FileStore
       def delete_all
-        FileUtils.rm_rf cache_path
+        FileUtils.rm_rf cache_path if File.exist?(cache_path)
       end
     end
   end
